Added particle manager

+++ PingEmDown/Randomizer/IRandomizer.cs
@@ -9,6 +9,6 @@ namespace PingEmDown.Randomizer
         int Next();
         int Next(int maxValue);
         int Next(int minValue, int maxValue);
-        double NextDouble();
+        float NextDouble();
     }
 }
